agent: break test_base_telem_classes into discrete test files
