+++ monkey/infection_monkey/telemetry/tests/test_exploit_telem.py
@@ -1,0 +1,31 @@
+import pytest
+
+from infection_monkey.exploit.wmiexec import WmiExploiter
+from infection_monkey.model.host import VictimHost
+from infection_monkey.telemetry.exploit_telem import ExploitTelem
+
+
+HOSTNAME = "hostname"
+DOMAIN_NAME = "domain-name"
+IP = "0.0.0.0"
+HOST = VictimHost(IP, DOMAIN_NAME)
+EXPLOITER = WmiExploiter(HOST)
+RESULT = False
+
+
+@pytest.fixture
+def exploit_telem_test_instance():
+    return ExploitTelem(EXPLOITER, RESULT)
+
+
+def test_exploit_telem_send(exploit_telem_test_instance, spy_send_telemetry):
+    exploit_telem_test_instance.send()
+    expected_data = {
+        "result": RESULT,
+        "machine": HOST.as_dict(),
+        "exploiter": EXPLOITER.__class__.__name__,
+        "info": EXPLOITER.exploit_info,
+        "attempts": EXPLOITER.exploit_attempts,
+    }
+    assert spy_send_telemetry.data == expected_data
+    assert spy_send_telemetry.telem_category == "exploit"
